Struct hyper_tls::HttpsConnector

pub struct HttpsConnector<T> { /* fields omitted */ }

A Connector for the https scheme.

Methods

impl HttpsConnector<HttpConnector>

pub fn new(threads: usize, handle: &Handle) -> Result<Self, Error>

Construct a new HttpsConnector.

Takes number of DNS worker threads.

This uses hyper's default HttpConnector, and default TlsConnector. If you wish to use something besides the defaults, use From::from.

impl<T> HttpsConnector<T> where
T: Connect,

pub fn danger_disable_hostname_verification(&mut self, disable: bool)

Disable hostname verification when connecting.

Warning

You should think very carefully before you use this method. If hostname verification is not used, any valid certificate for any site will be trusted for use from any other. This introduces a significant vulnerability to man-in-the-middle attacks.