Struct hyper_tls::HttpsConnector

pub struct HttpsConnector<T> { /* private fields */ }

A Connector for the https scheme.

Implementations

impl HttpsConnector<HttpConnector>

pub fn new() -> Self

Construct a new HttpsConnector.

This uses hyper’s default HttpConnector, and default TlsConnector. If you wish to use something besides the defaults, use From::from.

Note

By default this connector will use plain HTTP if the URL provded uses the HTTP scheme (eg: http://example.com/).

If you would like to force the use of HTTPS then call https_only(true) on the returned connector.

Panics

This will panic if the underlying TLS context could not be created.

To handle that error yourself, you can use the HttpsConnector::from constructor after trying to make a TlsConnector.

impl<T> HttpsConnector<T>

pub fn https_only(&mut self, enable: bool)

Force the use of HTTPS when connecting.

If a URL is not https when connecting, an error is returned.

pub fn new_with_connector(http: T) -> Self

With connector constructor